09-serverless/pyproject.toml
[project]
name = "09-serverless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "requests>=2.32.5",
]
